Half-quote all file paths
This works for paths that contain spaces and paths that end with a
backslash. If a trailing quote is used, a trailing backslash in the
argument will escape the quote, but if only the start of the argument is
quoted, the argument will continue until the end of the command line.