Silent install
An argument to the script which
a) don't require the user to press enter
b) doesn't open the control panel

I'm trying to build a Chocolatey package, complete silence and automation is required. I'm not the best at batch script, so if this PR isn't up to your standards, please implement this feature in your own way.