Use PowerShell to modify %PATH% and notify the shell
The main reason I chose batch over PowerShell for these scripts was that
batch files can be double-clicked by the user. Also, while batch is
restrictive and awful, PowerShell is clunky and overly verbose.

Despite this, PowerShell has some uses, and it can be invoked from cmd
without execution policy restrictions getting in the way. Manipulating
the %PATH% in a batch file would be difficult to do without accidentally
expanding embedded environment variables or trimming long PATHs, but
PowerShell allows this to be done easily and safely.

PowerShell can also be used to notify the shell of changes to the
environment variables and the file associations, so do that too.